rag-demo/requirements.txt
# RAG Demo System Dependencies
# Python 3.9+

# Core
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.0.5

# OpenAI (optional, for embedding and LLM)
openai>=1.6.0

# Text processing
markdown>=3.5.0
unstructured>=0.11.0
unstructured[md]>=0.11.0

# BM25 for keyword search
rank-bm25>=0.2.2

# Web API
fastapi>=0.109.0
uvicorn>=0.25.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
